requirements.txt
openai
pypdf
python-docx
gspread
google-auth
